requirements.txt
openai>=1.0.0
requests>=2.31.0
tenacity>=8.2.3
datasets>=2.18.0
pandas>=2.1.0
numpy>=1.26.0
tqdm>=4.66.0
pyyaml>=6.0.1
matplotlib>=3.8.0
scipy>=1.11.0
statsmodels>=0.14.0
sentence-transformers>=2.6.0
orjson>=3.9.0
socksio>=1.0.0
